refactor: 채팅 상태바 타입/스타일 중복 제거 및 목업 분리 (#280)
- ToolStatusIconType을 chat_types로 통합해 아이콘 유니온 타입 중복 제거
- 도구 상태바 목업 데이터/헬퍼를 tool_status_mock.ts로 분리
- 상태바 공통 컨테이너 클래스를 STATUS_BAR_BASE_CLASS 상수로 추출
  (ToolStatusBar / ThinkingBar / FinalResponseLoadingBar 공유)
