+++ src/features/chat/components/rightpanel/status_bar_styles.ts
@@ -1,0 +1,8 @@
+/**
+ * AI 상태바 공통 컨테이너 스타일
+ *
+ * ThinkingBar / FinalResponseLoadingBar / ToolStatusBar가 공유하는
+ * 둥근 회색 바 컨테이너의 기본 클래스. 패딩은 각 컴포넌트에서 덧붙인다.
+ */
+export const STATUS_BAR_BASE_CLASS =
+  "flex min-h-[30px] w-full items-center gap-[8px] overflow-hidden rounded-[12px] border-[0.5px] border-[#D1D6DE] bg-[#E3E7ED]";
